requirements.txt
fastapi==0.89.1
python-multipart==0.0.5
pydantic==1.10.4
uvicorn==0.20.0
gunicorn==20.1.0
requests==2.28.1
PyRSMQ==0.4.5
redis==4.4.2
httpx==0.23.3
